requirements.txt
streamlit
plotly
python-dotenv
pymongo
cloudinary
certifi